some controllers and models made

+++ db/migrate/20210329200326_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[6.1]
       t.string :username
       t.string :password
       t.integer :reward_points
-      t.integer :balance
+      t.float :balance
 
       t.timestamps
     end
